% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stovol.R
\name{stovol}
\alias{stovol}
\title{stovol}
\usage{
stovol(data, burn_in, N, model_type)
}
\arguments{
\item{data}{A data set object, which is one dimension time series}

\item{burn_in}{The sample size of the burn in before estimating the parameters}

\item{N}{Sample size of the totally sampled time series}

\item{model_type}{Specify the type of the model that will fitted.(BSV, ASV)}
}
\value{
(1) The estimated parameters and their confidence intervals,
(2) the sampled time series after a burn_in period for parameter estimation,
(3) the probability integral transform of the fitted model based on the training data,
(4) the standardized residuals , and (5) the AIC an BIC.
}
\description{
Fits two types of stochasitc volatility models
}
\examples{
stovol(data, burn_in, N, model_type)
}
