% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{ts_plot}
\alias{ts_plot}
\title{stovol}
\usage{
ts_plot(est)
}
\arguments{
\item{est}{A datafram of the time series}
}
\description{
plot the time series draw from the conditional distributions.
}
\examples{
ts_plot(est)
}
