% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_plot.R
\name{vol_plot}
\alias{vol_plot}
\title{stovol}
\usage{
vol_plot(y, z)
}
\arguments{
\item{y}{the training time series}

\item{z}{The estimated volatility}
}
\description{
compare the absolute values of the returns with the estimated volatilizes
}
\examples{
vol_plot(z)
}
